requirements.txt
pandas
google-cloud-secret-manager
gcloud-config-helper
setuptools
google-cloud-bigquery
db-dtypes